# perf: avoid building module graph if dynamic specifier already in graph