--- a/tests/specs/run/dynamic_already_prepared/dynamic.ts
+++ b/tests/specs/run/dynamic_already_prepared/dynamic.ts
@@ -1,0 +1,1 @@
+console.log(1);