--- a/tests/specs/run/dynamic_already_prepared/dynamic2.ts
+++ b/tests/specs/run/dynamic_already_prepared/dynamic2.ts
@@ -1,0 +1,1 @@
+await import("./dynamic3.ts");