--- a/tests/specs/run/dynamic_already_prepared/dynamic3.ts
+++ b/tests/specs/run/dynamic_already_prepared/dynamic3.ts
@@ -1,0 +1,1 @@
+await import("./dynamic.ts");