--- a/tests/specs/run/dynamic_already_prepared/main.out
+++ b/tests/specs/run/dynamic_already_prepared/main.out
@@ -1,0 +1,1 @@
+[WILDCARD]Prepared module load.[WILDCARD]Skipping prepare module load.[WILDCARD]Skipping prepare module load.[WILDCARD]Skipping prepare module load.[WILDCARD]