--- a/tests/specs/run/dynamic_already_prepared/main.ts
+++ b/tests/specs/run/dynamic_already_prepared/main.ts
@@ -1,0 +1,2 @@
+await import("./dynamic.ts");
+await import("./dynamic2.ts");